requirements.txt
keras==2.12.0
matplotlib==3.5.2
numpy>=1.21.5
opencv_python==4.7.0.68
pandas==1.4.4
Pillow==9.4.0
plotly==5.9.0
requests==2.28.1
scikit_learn==1.2.2
seaborn==0.11.2
streamlit==1.20.0
tensorflow==2.12.0
tensorflow_datasets==4.8.2
torch==2.0.0
torchvision==0.15.1